Rename from sql to sqlclient
