+++ src/SQLClient/getConnectionParams.js
